Add tests for SHA checksums
And move tests to own files, add clippy run to CI
